requirements.txt
numpy==1.23.5
pillow==10.4.0
opencv-python==4.11.0
scikit-learn==1.3.2
tensorflow==2.12.0
keras==2.12.0
face_recognition==1.2.3